CodeAware/backend/requirements.txt
# Core API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.1

# Caching & Queue
redis==5.0.1
celery==5.3.4
flower==2.0.1

# Code Analysis
ast-parser==1.0.0
radon==6.0.1
bandit==1.7.5
pylint==3.0.3
flake8==6.1.0
mypy==1.7.1
semgrep==1.45.0
safety==2.3.5
vulture==2.10
lizard==1.17.10

# Machine Learning
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.3
joblib==1.3.2

# Git Integration
gitpython==3.1.40
PyGithub==2.1.1
python-gitlab==4.2.0

# Utilities
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1
python-dateutil==2.8.2
pytz==2023.3

# Monitoring & Logging
prometheus-client==0.19.0
sentry-sdk==1.38.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0




